// kestrel/src/types/engine/Auth.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ApiKeyLocation } from "./ApiKeyLocation";

export type Auth =
  | { type: "none" }
  | { type: "bearer"; token: string }
  | { type: "basic"; username: string; password: string }
  | { type: "apiKey"; location: ApiKeyLocation; name: string; value: string };