// kestrel/src/types/engine/Body.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FormField } from "./FormField";
import type { KeyValue } from "./KeyValue";

export type Body = { "type": "none" } | { "type": "json", content: string, } | { "type": "raw", contentType: string, content: string, } | { "type": "form", fields: Array<KeyValue>, } | { "type": "multipart", fields: Array<FormField>, };